Fixed no subscribers KeyError bug
Formerly, when events were dispatched to a group that contains at
least one subscriber, if there were no subscribers for a dispatched
event, a KeyError would be raised. This has been fixed so that in all
cases, it's Okay to dispatch events where there are no subscribers to
handle them. (doh!)