broadcast and broadcast_default
The broadcast and broadcast_default methods dispatch events to
listeners immediately in a blocking operation. The events are never
inserted into the queue and so they will not double dispatch when
publish is eventually called.

This is useful in situations where the publish/dispatch cycle is too
complex or direct control over dispatch sequencing is required.